remove old EventBuffer type
