+++ lib/saluki-core/src/topology/interconnect/event_buffer.rs
@@ -2,36 +2,65 @@ use std::{collections::VecDeque, fmt};
 
 use saluki_event::{DataType, Event};
 
-use crate::pooling::helpers::pooled;
+use crate::pooling::{helpers::pooled_newtype, Clearable};
 
-pooled! {
-    /// An event buffer.
-    ///
-    /// Event buffers are able to hold an arbitrary number of [`Event`]s within them, and are reclaimed by their
-    /// originating buffer pool once dropped. This allows for efficient reuse of the event buffers as the underlying
-    /// allocations to hold a large number of events can itself grow large.
-    struct EventBuffer {
-        events: VecDeque<Event>,
-        seen_data_types: DataType,
-    }
+/// A double-ended queue implemented with a ring buffer that has a fixed capacity at creation.
+struct FixedSizeVecDeque<T>(VecDeque<T>);
 
-    clear => |this| { this.events.clear(); this.seen_data_types = DataType::none(); }
+/// A fixed-size event buffer.
+pub struct FixedSizeEventBufferInner {
+    events: FixedSizeVecDeque<Event>,
+    seen_data_types: DataType,
 }
 
-impl EventBuffer {
-    /// Returns the total number of events the event buffer can hold without reallocating.
-    pub fn capacity(&self) -> usize {
-        self.data().events.capacity()
+impl FixedSizeEventBufferInner {
+    /// Creates a new fixed-size event buffer with the given capacity.
+    pub fn with_capacity(capacity: usize) -> Self {
+        Self {
+            events: FixedSizeVecDeque(VecDeque::with_capacity(capacity)),
+            seen_data_types: DataType::none(),
+        }
+    }
+}
+
+impl Clearable for FixedSizeEventBufferInner {
+    fn clear(&mut self) {
+        self.events.0.clear();
+        self.seen_data_types = DataType::none();
+    }
+}
+
+pooled_newtype! {
+    outer => FixedSizeEventBuffer,
+    inner => FixedSizeEventBufferInner,
+}
+
+impl FixedSizeEventBuffer {
+    #[cfg(test)]
+    pub fn for_test(capacity: usize) -> Self {
+        use crate::pooling::helpers::get_pooled_object_via_builder;
+
+        get_pooled_object_via_builder(|| FixedSizeEventBufferInner::with_capacity(capacity))
     }
 
-    /// Returns `true` if the event buffer contains no events.
-    pub fn is_empty(&self) -> bool {
-        self.data().events.is_empty()
+    /// Returns the total number of events the event buffer can hold.
+    pub fn capacity(&self) -> usize {
+        self.data().events.0.capacity()
     }
 
     /// Returns the number of events in the event buffer.
     pub fn len(&self) -> usize {
-        self.data().events.len()
+        self.data().events.0.len()
+    }
+
+    /// Returns `true` if the event buffer contains no events.
+    pub fn is_empty(&self) -> bool {
+        self.data().events.0.is_empty()
+    }
+
+    /// Returns `true` if the event buffer has no remaining capacity.
+    pub fn is_full(&self) -> bool {
+        self.len() == self.capacity()
     }
 
     /// Returns `true` if this event buffer contains one or more events of the given data type.
@@ -39,21 +68,17 @@ impl EventBuffer {
         self.data().seen_data_types.contains(data_type)
     }
 
-    /// Appends an event to the back of the event buffer.
-    pub fn push(&mut self, event: Event) {
-        self.data_mut().seen_data_types |= event.data_type();
-        self.data_mut().events.push_back(event);
-    }
+    /// Attempts to append an event to the back of the event buffer.
+    ///
+    /// If the event buffer is full, `Some` is returned with the original event.
+    pub fn try_push(&mut self, event: Event) -> Option<Event> {
+        if self.len() == self.capacity() {
+            return Some(event);
+        }
 
-    /// Reserves capacity for at least `additional` more elements to be inserted in the event buffer.
-    ///
-    /// The event buffer may reserve more space to speculatively avoid frequent reallocations.
-    ///
-    /// ## Panics
-    ///
-    /// Panics if the new capacity overflows `usize`.
-    pub fn reserve(&mut self, additional: usize) {
-        self.data_mut().events.reserve(additional);
+        self.data_mut().seen_data_types |= event.data_type();
+        self.data_mut().events.0.push_back(event);
+        None
     }
 
     /// Extract events from the event buffer given a predicate function.
@@ -66,7 +91,7 @@ impl EventBuffer {
         let mut indices_to_remove = Vec::new();
         let mut removed_events = VecDeque::new();
 
-        for (pos, event) in events.iter_mut().enumerate() {
+        for (pos, event) in events.0.iter_mut().enumerate() {
             if predicate(event) {
                 indices_to_remove.push(pos);
             }
@@ -74,8 +99,8 @@ impl EventBuffer {
 
         // Remove elements from back to front to avoid index shifting issues
         for &pos in indices_to_remove.iter().rev() {
-            if pos < events.len() {
-                removed_events.push_back(events.swap_remove_back(pos).unwrap());
+            if pos < events.0.len() {
+                removed_events.push_back(events.0.swap_remove_back(pos).unwrap());
             }
         }
 
@@ -83,26 +108,15 @@ impl EventBuffer {
     }
 }
 
-impl fmt::Debug for EventBuffer {
+impl fmt::Debug for FixedSizeEventBuffer {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("EventBuffer")
-            .field("event_len", &self.data().events.len())
+        f.debug_struct("FixedSizeEventBuffer")
+            .field("event_len", &self.len())
             .finish()
     }
 }
 
-impl Extend<Event> for EventBuffer {
-    fn extend<T>(&mut self, iter: T)
-    where
-        T: IntoIterator<Item = Event>,
-    {
-        for event in iter {
-            self.push(event);
-        }
-    }
-}
-
-impl IntoIterator for EventBuffer {
+impl IntoIterator for FixedSizeEventBuffer {
     type Item = Event;
     type IntoIter = IntoIter;
 
@@ -111,33 +125,33 @@ impl IntoIterator for EventBuffer {
     }
 }
 
-impl<'a> IntoIterator for &'a EventBuffer {
+impl<'a> IntoIterator for &'a FixedSizeEventBuffer {
     type Item = &'a Event;
     type IntoIter = std::collections::vec_deque::Iter<'a, Event>;
 
     fn into_iter(self) -> Self::IntoIter {
-        self.data().events.iter()
+        self.data().events.0.iter()
     }
 }
 
-impl<'a> IntoIterator for &'a mut EventBuffer {
+impl<'a> IntoIterator for &'a mut FixedSizeEventBuffer {
     type Item = &'a mut Event;
     type IntoIter = std::collections::vec_deque::IterMut<'a, Event>;
 
     fn into_iter(self) -> Self::IntoIter {
-        self.data_mut().events.iter_mut()
+        self.data_mut().events.0.iter_mut()
     }
 }
 
 pub struct IntoIter {
-    inner: EventBuffer,
+    inner: FixedSizeEventBuffer,
 }
 
 impl Iterator for IntoIter {
     type Item = Event;
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.inner.data_mut().events.pop_front()
+        self.inner.data_mut().events.0.pop_front()
     }
 }
 
@@ -152,20 +166,32 @@ mod tests {
         DataType, Event,
     };
 
-    use super::EventBuffer;
-    use crate::pooling::{helpers::get_pooled_object_via_default, Clearable as _};
+    use super::{FixedSizeEventBuffer, FixedSizeEventBufferInner};
+    use crate::pooling::{helpers::get_pooled_object_via_builder, Clearable as _};
+
+    #[test]
+    fn capacity() {
+        let mut buffer =
+            get_pooled_object_via_builder::<_, FixedSizeEventBuffer>(|| FixedSizeEventBufferInner::with_capacity(2));
+        assert_eq!(buffer.capacity(), 2);
+
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 42.0))).is_none());
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 43.0))).is_none());
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 44.0))).is_some());
+    }
 
     #[test]
     fn clear() {
         // Create an empty event buffer and assert that it's empty and has no seen data types:
-        let mut buffer = get_pooled_object_via_default::<EventBuffer>();
+        let mut buffer =
+            get_pooled_object_via_builder::<_, FixedSizeEventBuffer>(|| FixedSizeEventBufferInner::with_capacity(10));
         assert!(buffer.is_empty());
         assert!(!buffer.has_data_type(DataType::Metric));
         assert!(!buffer.has_data_type(DataType::EventD));
         assert!(!buffer.has_data_type(DataType::ServiceCheck));
 
         // Now write a metric, and make sure that's reflected:
-        buffer.push(Event::Metric(Metric::counter("foo", 42.0)));
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 42.0))).is_none());
         assert!(!buffer.is_empty());
         assert!(buffer.has_data_type(DataType::Metric));
         assert!(!buffer.has_data_type(DataType::EventD));
@@ -182,17 +208,18 @@ mod tests {
 
     #[test]
     fn has_data_type() {
-        let mut buffer = get_pooled_object_via_default::<EventBuffer>();
+        let mut buffer =
+            get_pooled_object_via_builder::<_, FixedSizeEventBuffer>(|| FixedSizeEventBufferInner::with_capacity(10));
         assert!(!buffer.has_data_type(DataType::Metric));
         assert!(!buffer.has_data_type(DataType::EventD));
         assert!(!buffer.has_data_type(DataType::ServiceCheck));
 
-        buffer.push(Event::Metric(Metric::counter("foo", 42.0)));
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 42.0))).is_none());
         assert!(buffer.has_data_type(DataType::Metric));
         assert!(!buffer.has_data_type(DataType::EventD));
         assert!(!buffer.has_data_type(DataType::ServiceCheck));
 
-        buffer.push(Event::EventD(EventD::new("title", "text")));
+        assert!(buffer.try_push(Event::EventD(EventD::new("title", "text"))).is_none());
         assert!(buffer.has_data_type(DataType::Metric));
         assert!(buffer.has_data_type(DataType::EventD));
         assert!(!buffer.has_data_type(DataType::ServiceCheck));
@@ -200,14 +227,19 @@ mod tests {
 
     #[test]
     fn extract() {
-        let mut buffer = get_pooled_object_via_default::<EventBuffer>();
-        buffer.push(Event::Metric(Metric::counter("foo", 42.0)));
-        buffer.push(Event::Metric(Metric::counter("foo", 43.0)));
-        buffer.push(Event::EventD(EventD::new("foo1", "bar1")));
-        buffer.push(Event::EventD(EventD::new("foo2", "bar2")));
-        buffer.push(Event::EventD(EventD::new("foo3", "bar3")));
-        buffer.push(Event::ServiceCheck(ServiceCheck::new("foo4", CheckStatus::Ok)));
-        buffer.push(Event::ServiceCheck(ServiceCheck::new("foo5", CheckStatus::Ok)));
+        let mut buffer =
+            get_pooled_object_via_builder::<_, FixedSizeEventBuffer>(|| FixedSizeEventBufferInner::with_capacity(10));
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 42.0))).is_none());
+        assert!(buffer.try_push(Event::Metric(Metric::counter("foo", 43.0))).is_none());
+        assert!(buffer.try_push(Event::EventD(EventD::new("foo1", "bar1"))).is_none());
+        assert!(buffer.try_push(Event::EventD(EventD::new("foo2", "bar2"))).is_none());
+        assert!(buffer.try_push(Event::EventD(EventD::new("foo3", "bar3"))).is_none());
+        assert!(buffer
+            .try_push(Event::ServiceCheck(ServiceCheck::new("foo4", CheckStatus::Ok)))
+            .is_none());
+        assert!(buffer
+            .try_push(Event::ServiceCheck(ServiceCheck::new("foo5", CheckStatus::Ok)))
+            .is_none());
 
         assert_eq!(buffer.len(), 7);
