+++ lib/saluki-core/src/topology/interconnect/mod.rs
@@ -1,13 +1,10 @@
 //! Component interconnects.
 
 mod event_buffer;
-pub use self::event_buffer::EventBuffer;
+pub use self::event_buffer::{FixedSizeEventBuffer, FixedSizeEventBufferInner};
 
 mod event_stream;
 pub use self::event_stream::EventStream;
-
-mod fixed_event_buffer;
-pub use self::fixed_event_buffer::{FixedSizeEventBuffer, FixedSizeEventBufferInner};
 
 mod forwarder;
 pub use self::forwarder::{BufferedForwarder, Forwarder};
